--- src/dialogs/COINiDNotFound/index.js
@@ -1,2 +1,0 @@
-import COINiDNotFound from './COINiDNotFound';
-export default COINiDNotFound;
