--- src/dialogs/QRDataSender/index.js
@@ -1,2 +1,0 @@
-import QRDataSender from './QRDataSender';
-export default QRDataSender;
