--- src/dialogs/SelectAddressType/index.js
@@ -1,2 +1,0 @@
-import SelectAddressType from './SelectAddressType';
-export default SelectAddressType;
