+++ src/dialogs/SelectColdTransportType.js
@@ -1,16 +1,78 @@
-
-
 import React, { PureComponent } from 'react';
 import PropTypes from 'prop-types';
-import { View } from 'react-native';
+import { StyleSheet, View } from 'react-native';
 import bleCentral from 'react-native-p2p-transfer-ble-central';
 
 import {
   DetailsModal, Text, Button, CheckBoxSelect,
-} from '../../components';
-import { fontWeight } from '../../config/styling';
-import settings from '../../config/settings';
-import styles from './styles';
+} from '../components';
+import { colors, fontWeight } from '../config/styling';
+import settings from '../config/settings';
+
+const styles = StyleSheet.create({
+  container: {
+    padding: 16,
+    paddingTop: 0,
+    maxHeight: 400,
+  },
+  loader: {
+    marginBottom: 40,
+    marginTop: 30,
+  },
+  amountText: {
+    marginBottom: 4,
+    textAlign: 'center',
+    ...fontWeight.normal,
+  },
+  fiatText: {
+    color: colors.gray,
+    fontSize: 24,
+    marginBottom: 16,
+    textAlign: 'center',
+    ...fontWeight.book,
+  },
+  outgoing: {
+    color: colors.orange,
+  },
+  incoming: {
+    color: colors.green,
+  },
+  footer: {
+    borderTopWidth: 1,
+    borderTopColor: colors.lightGray,
+  },
+  footerBtn: {
+    alignItems: 'center',
+    flexDirection: 'row',
+    justifyContent: 'center',
+    paddingTop: 16,
+    paddingBottom: 16,
+  },
+  footerLink: {
+    fontSize: 16,
+    marginRight: 5,
+    textAlign: 'center',
+  },
+
+  row: {
+    alignItems: 'stretch',
+    flexDirection: 'row',
+  },
+  rowTitle: {
+    color: colors.gray,
+    marginTop: 8,
+  },
+  rowContainer: {
+    flex: 1,
+  },
+  rowData: {
+    flex: 1,
+    textAlign: 'right',
+  },
+  rowText: {
+    fontSize: 16,
+  },
+});
 
 export default class SelectColdTransportType extends PureComponent {
   constructor(props) {
@@ -32,12 +94,11 @@ export default class SelectColdTransportType extends PureComponent {
   }
 
   componentDidMount() {
-    bleCentral.isSupported()
-      .then((isBLESupported) => {
-        this.setState({
-          isBLESupported,
-        });
+    bleCentral.isSupported().then((isBLESupported) => {
+      this.setState({
+        isBLESupported,
       });
+    });
   }
 
   _open = (onSelectCb) => {
@@ -95,7 +156,9 @@ export default class SelectColdTransportType extends PureComponent {
           </Text>
           <View style={{ marginTop: 16 }}>
             <CheckBoxSelect
-              onIndexChange={(newIndex) => { this.setState({ selectedIndex: newIndex }); }}
+              onIndexChange={(newIndex) => {
+                this.setState({ selectedIndex: newIndex });
+              }}
               selectedIndex={selectedIndex}
               data={selectData}
             />
