--- src/dialogs/SelectColdTransportType/index.js
@@ -1,2 +1,0 @@
-import SelectColdTransportType from './SelectColdTransportType';
-export default SelectColdTransportType;
