--- src/dialogs/SetupWallet/index.js
@@ -1,2 +1,0 @@
-import SetupWallet from './SetupWallet';
-export default SetupWallet;
