--- src/dialogs/SweepPrivateKey/index.js
@@ -1,3 +1,0 @@
-import SweepPrivateKey from './SweepPrivateKey';
-
-export default SweepPrivateKey;
