--- src/dialogs/ValidateAddress/index.js
@@ -1,2 +1,0 @@
-import ValidateAddress from './ValidateAddress';
-export default ValidateAddress;
